refactor: used code style
