+++ src/Tochka.JsonRpc.Client/IJsonRpcClient.cs
@@ -2,8 +2,8 @@
 using System.Text.Json;
 using JetBrains.Annotations;
 using Tochka.JsonRpc.Client.Models;
-using Tochka.JsonRpc.Client.Models.Batch;
-using Tochka.JsonRpc.Client.Models.Single;
+using Tochka.JsonRpc.Client.Models.BatchResult;
+using Tochka.JsonRpc.Client.Models.SingleResult;
 using Tochka.JsonRpc.Common.Models.Id;
 using Tochka.JsonRpc.Common.Models.Request;
 
@@ -155,7 +155,7 @@ public interface IJsonRpcClient
     /// <exception cref="System.ArgumentException">When requestUrl starts with '/'</exception>
     Task<ISingleJsonRpcResult> SendRequest<TParams>(IRpcId id, string method, TParams? parameters, CancellationToken cancellationToken)
         where TParams : class;
-    
+
     /// <summary>
     /// Send request to given url. Expects HTTP 200 with JSON-RPC Rpc typed response
     /// </summary>
@@ -217,12 +217,12 @@ public interface IJsonRpcClient
     Task<ISingleJsonRpcResult<TResponse>> SendRequest<TParams, TResponse>(string method, TParams? parameters, CancellationToken cancellationToken)
         where TParams : class
         where TResponse : class;
-    
+
     /// <summary>
     /// Send request to given url. Expects HTTP 200 with JSON-RPC typed response
     /// </summary>
     /// <typeparam name="TParams">Type of params</typeparam>
-    /// <typeparam name="TResponse">Type of response</typeparam> 
+    /// <typeparam name="TResponse">Type of response</typeparam>
     /// <param name="requestUrl">Relative path, appended to BaseAddress. Must not start with '/'</param>
     /// <param name="id">JSON-RPC request id. Can be null</param>
     /// <param name="method">JSON-RPC method</param>
@@ -240,7 +240,7 @@ public interface IJsonRpcClient
     /// Send request to BaseUrl. Expects HTTP 200 with JSON-RPC typed response
     /// </summary>
     /// <typeparam name="TParams">Type of params</typeparam>
-    /// <typeparam name="TResponse">Type of response</typeparam> 
+    /// <typeparam name="TResponse">Type of response</typeparam>
     /// <param name="id">JSON-RPC request id. Can be null</param>
     /// <param name="method">JSON-RPC method</param>
     /// <param name="parameters">JSON-RPC params - This member MAY be omitted</param>
@@ -252,7 +252,7 @@ public interface IJsonRpcClient
     Task<ISingleJsonRpcResult<TResponse>> SendRequest<TParams, TResponse>(IRpcId id, string method, TParams? parameters, CancellationToken cancellationToken)
         where TParams : class
         where TResponse : class;
-    
+
     /// <summary>
     /// Send batch of requests or notifications to given url. Expects HTTP 200 with batch JSON-RPC response if batch contains at least one request
     /// </summary>
@@ -282,7 +282,7 @@ public interface IJsonRpcClient
     /// <param name="requestUrl">Relative path, appended to BaseAddress. Must not start with '/'</param>
     /// <param name="calls">JSON-RPC requests or notifications</param>
     /// <param name="cancellationToken"></param>
-    /// <typeparam name="TResponse">Type of response</typeparam> 
+    /// <typeparam name="TResponse">Type of response</typeparam>
     /// <returns>Result to be inspected for response data or errors</returns>
     /// <exception cref="JsonRpcException">When HTTP status code is not 200, body deserialized as single response, response count does not match requests in batch</exception>
     /// <exception cref="JsonException">When reading or deserializing JSON from body failed</exception>
@@ -294,7 +294,7 @@ public interface IJsonRpcClient
     /// </summary>
     /// <param name="calls">JSON-RPC requests or notifications</param>
     /// <param name="cancellationToken"></param>
-    /// <typeparam name="TResponse">Type of response</typeparam> 
+    /// <typeparam name="TResponse">Type of response</typeparam>
     /// <returns>Result to be inspected for response data or errors</returns>
     /// <exception cref="JsonRpcException">When HTTP status code is not 200, body deserialized as single response, response count does not match requests in batch</exception>
     /// <exception cref="JsonException">When reading or deserializing JSON from body failed</exception>
